Moved SearchSpace and convert_params to separate file
